refactor: 💡 Example now using webpack
