+++ example/package.json
@@ -4,16 +4,25 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "parcel src/index.html src/**/*.html"
+    "start": "webpack-dev-server",
+    "build": "webpack"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "parcel": "^1.10.3"
+    "@biotope/element": "^3"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3"
+    "babel-core": "^6.26.3",
+    "glob": "^7.1.3",
+    "html-webpack-plugin": "^4.0.0-beta.5",
+    "ts-loader": "^5.3.3",
+    "typescript": "^3.3.3333",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.2.3",
+    "webpack-dev-server": "^3.2.1",
+    "webpack-glob-entry": "^2.1.1"
   },
   "browserslist": [
     "last 1 chrome versions"
