+++ example/src/index.html
@@ -10,28 +10,31 @@
       .example {
         display: block;
       }
+
     </style>
   </head>
 
   <body>
     <h1>BioElement Demos</h1>
-    <a class="example" href="/simple-ts/simple.html">Typescript: Basic Component</a>
-    
-    <a class="example" href="/nested-ts/nested.html">Typescript: Nested components</a>
-    <a class="example" href="/nested-js/nested.html">Javascript: Nested components</a>
-    
-    <a class="example" href="/slots-ts/slots.html">Typescript: Shadow Root Slots</a>
-    <a class="example" href="/nested-slots-ts/nested-slots.html">Typescript: Nested Slots</a>
-    <a class="example" href="/two-slots-ts/slots.html">Typescript: Two slots</a>
-    
-    <a class="example" href="/styles-js/styles.html">Javascript: Styles</a>
-    <a class="example" href="/events-js/events.html">Javascript: Events</a>
-    <a class="example" href="/button-js/button.html">Javascript: Button</a>
-    
-    <a class="example" href="/state-from-props-js/state-from-props.html">Javascript: State from Props</a>
-    <a class="example" href="/set-focus-js/set-focus.html">Javascript: Set focus</a>
+    <h2>Javascript Demos</h2>
+    <a class="example" href="/js/button/index.html">Button</a>
+    <a class="example" href="/js/nested/index.html">Nested components</a>
+    <a class="example" href="/js/styles/index.html">Styles</a>
+    <a class="example" href="/js/events/index.html">Events</a>
+    <a class="example" href="/js/state-from-props/index.html">State from Props</a>
+    <a class="example" href="/js/set-focus/index.html">Set focus</a>
+
+    <h2>Typescript Demos</h2>
+    <a class="example" href="/ts/simple/index.html">Basic Component</a>
+    <a class="example" href="/ts/nested/index.html">Nested components</a>
+    <a class="example" href="/ts/slots/index.html">Shadow Root Slots</a>
+    <a class="example" href="/ts/nested-slots/index.html">Nested Slots</a>
+    <a class="example" href="/ts/two-slots/index.html">Two slots</a>
+
+
 
     <p>⚠️ These examples are not polyfilled, we've an open <a href="https://github.com/biotope/biotope-element/issues/59">issue</a>
       and we'll hopefully fix that issue soonish.</p>
   </body>
+
 </html>
