+++ example/src/js/nested/child.js
@@ -1,0 +1,14 @@
+import Component from '@biotope/element';
+
+export class TextInput extends Component {
+  render() {
+    this.html`
+      <label>Label</label>
+      <input type="text">
+    `;
+  }
+}
+
+TextInput.componentName = 'x-text-input';
+
+TextInput.register();
