--- example/src/nested-slots-ts/nested-slots.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-	<head>
-		<meta charset="UTF-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<meta http-equiv="X-UA-Compatible" content="ie=edge">
-		<title>BioElement</title>
-		<script src="./index.ts"></script>
-	</head>
-
-	<body>
-		<x-slots>
-			<span>I will be nested slotted</span>
-		</x-slots>
-	</body>
-
-</html>
