--- example/src/nested-ts/nested.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>BioElement</title>
-  <script src="./index.ts"></script>
-</head>
-
-<body>
-  <x-form></x-form>
-</body>
-
-</html>
