--- example/src/slots-ts/slots.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>BioElement</title>
-  <script src="./index.ts"></script>
-</head>
-
-<body>
-
-
-  <x-slots>
-    <p>I am slotted.</p>
-  </x-slots>
-
-</body>
-
-</html>
