+++ example/src/ts/nested-slots/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+
+	<head>
+		<meta charset="UTF-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<meta http-equiv="X-UA-Compatible" content="ie=edge">
+		<title>BioElement</title>
+	</head>
+
+	<body>
+		<x-slots>
+			<span>I will be nested slotted</span>
+		</x-slots>
+	</body>
+
+</html>
