+++ example/src/ts/nested-slots/index.ts
@@ -1,0 +1,16 @@
+import Component from '@biotope/element';
+import ChildSlot from './ChildSlot';
+
+class Slots extends Component<any, any> {
+  static componentName = 'x-slots';
+  static dependencies = [ChildSlot]
+  render() {
+    this.html`
+    <x-child-slot>
+      <slot />
+    </x-child-slot>
+    `;
+  }
+}
+
+Slots.register();
