+++ example/src/ts/two-slots/index.ts
@@ -1,0 +1,20 @@
+import Component from '@biotope/element';
+
+class FullName extends Component<any, any> {
+  static componentName = 'x-full-name';
+  render() {
+    this.html`
+      <style>
+        ::slotted(*) {
+          color: orange;
+        }
+      </style>
+      Firstname:
+      <slot name="first">FIRSTNAME</slot>
+      Lastname:
+      <slot name="last">LASTNAME</slot>
+    `;
+  }
+}
+
+FullName.register();
