--- example/src/two-slots-ts/index.ts
@@ -1,20 +1,0 @@
-import Component from '../../../src/index';
-
-class FullName extends Component<any, any> {
-  static componentName = 'x-full-name';
-  render() {
-    this.html`
-      <style>
-        ::slotted(*) {
-          color: orange;
-        }
-      </style>
-      Firstname:
-      <slot name="first">FIRSTNAME</slot>
-      Lastname:
-      <slot name="last">LASTNAME</slot>
-    `;
-  }
-}
-
-FullName.register();
